requirements.txt
streamlit
python-dotenv
groq
spacy
pymupdf
python-docx
langdetect
fpdf
